requirements.txt
django
Pillow
django-camera-imagefield